illuminate/requirements.txt
langchain
langchain-community
pypdf
openai
sievedata
tiktoken
pydub
